Dodanie logowania, rejestracji. Usuniecie mozliwosci usuniecia i edycji notatek innych uzytkownikow

+++ base/views.py
@@ -1,8 +1,58 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Note, Topic
 from .forms import NoteForm
 # Create your views here.
 
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method =="POST":
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request,username=username, password=password)
+        
+        if user is not None:
+            login(request,  user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or Password does not exist')
+
+    context = {'page':page}
+    return render(request, 'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occurred during registration')
+
+    return render(request,'base/login_register.html',{'form' : form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -18,6 +68,7 @@ def note(request, pk):
     context = {'note': note}
     return render(request, 'base/note.html',context)
 
+@login_required(login_url='login')
 def createNote(request):
     form = NoteForm()
     if request.method =='POST':
@@ -28,9 +79,13 @@ def createNote(request):
     context = {'form': form}
     return render(request, 'base/note_form.html', context)
 
+@login_required(login_url='login')
 def updateNote(request,pk):
     note = Note.objects.get(id = pk)
     form = NoteForm(instance=note)
+
+    if request.user != note.host:
+        return HttpResponse('You are not allowed here')
 
     if request.method == 'POST':
         form = NoteForm(request.POST, instance=note)
@@ -41,8 +96,13 @@ def updateNote(request,pk):
     context = {'form': form}
     return render(request, 'base/note_form.html', context)
 
+@login_required(login_url='login')
 def deleteNote(request,pk):
     note = Note.objects.get(id=pk)
+
+    if request.user != note.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method =="POST":
         note.delete()
         return redirect('home')
